requirements/base.txt
Django==3.0.7
djangorestframework==3.11.0
gunicorn==19.9.0
psycopg2==2.7.3.2
PyDictionary==2.0.1
wordcloud==1.8.1
numpy==1.19.5
